+++ NEWS.md
@@ -1,5 +1,10 @@
 # workflows (development version)
 
+* `workflow()` has gained new `preprocessor` and `spec` arguments for adding
+  a preprocessor (such as a recipe or formula) and a parsnip model specification
+  directly to a workflow upon creation. In many cases, this can reduce the
+  lines of code required to construct a complete workflow (#108).
+  
 * New `extract_*()` functions have been added that supersede the existing
   `pull_*()` functions. This is part of a larger move across the tidymodels
   packages towards a family of generic `extract_*()` functions. The `pull_*()`